# services/competition-service/openapi/competition-service.v1.yaml
openapi: 3.0.3
info:
  title: competition-service API
  version: 0.4.0
  description: |
    Dịch vụ Thi đấu của Badminton Digital — chấm trình, điểm trình, bảng xếp hạng, xếp cặp.
    Bước 1: người chơi, chấm trình, BXH trình độ, thuật toán xếp cặp. Bước 2: trận đấu, giải đấu, BXH thành tích, thống kê.
    Bước 3: buổi giao lưu (điểm danh, xếp sân trống, màn hình lớn, đóng buổi).
    Plan 19: bấm điểm trực tiếp từng pha cầu, luồng SSE đẩy tỉ số lên màn hình TV.

    Xác thực: service token ES256 (gateway của app chính ký, sống ≤ 5 phút) — xem docs/01-kien-truc.md mục 6.
    Envelope: { success, data, message, errors } giống app chính, thêm `code` khi lỗi.
    Mọi POST có tác dụng phụ bắt buộc header Idempotency-Key.
servers:
  - url: /
security:
  - serviceToken: []
...
paths:
...
  /v1/me/tournaments:
    get:
      tags: [tournament]
      summary: Giải của tôi
      description: 'Scope: rating:self'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [items]
                        properties:
                          items:
                            type: array
                            items:
                              type: object
                              required: [tournament, entryStatus, teamId, partnerPlayerId, upcomingMatches, playedMatches, placement]
                              properties:
                                tournament: { $ref: '#/components/schemas/Tournament' }
                                entryStatus: { type: string }
                                teamId: { type: string, nullable: true }
                                partnerPlayerId: { type: string, nullable: true }
                                upcomingMatches: { type: integer }
                                playedMatches: { type: integer }
                                placement:
                                  type: object
                                  nullable: true
                                  properties:
                                    from: { type: integer }
                                    to: { type: integer }
                                    label: { type: string }
        default: { $ref: '#/components/responses/Error' }
...
  /v1/tournaments/advice:
    post:
      tags: [tournament]
      summary: Gợi ý thể thức / số bảng / ước tính số trận và thời gian
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [teams]
              properties:
                teams: { type: integer, minimum: 2, maximum: 256 }
                courts: { type: integer, minimum: 1, maximum: 64 }
                matchMinutes: { type: integer, minimum: 1, maximum: 240 }
                format: { type: string, enum: [round_robin, groups_knockout, knockout] }
                groupCount: { type: integer, minimum: 2, maximum: 64 }
                advancePerGroup: { type: integer, enum: [1, 2] }
                thirdPlace: { type: boolean }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [format, suggestedFormat, estimate]
                        properties:
                          format: { type: string }
                          suggestedFormat: { type: string }
                          groupCount: { type: integer, nullable: true }
                          groupSizes: { type: array, nullable: true, items: { type: integer } }
                          advancePerGroup: { type: integer, nullable: true }
                          estimate:
                            type: object
                            required: [matches, slots, minutes]
                            properties:
                              matches: { type: integer }
                              slots: { type: integer }
                              minutes: { type: integer }
        default: { $ref: '#/components/responses/Error' }
  /v1/tournaments:
    post:
      tags: [tournament]
      summary: Tạo giải (nháp)
      description: 'Scope: tournament:manage. organizerRef phải thuộc claim org; tier chain cần org "*".'
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TournamentInput' }
      responses:
        '201': { $ref: '#/components/responses/TournamentDetail' }
        default: { $ref: '#/components/responses/Error' }
    get:
      tags: [tournament]
      summary: Danh sách giải (trong phạm vi chi nhánh của token)
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [draft, open, drawn, in_progress, finalized, cancelled] } }
        - { name: discipline, in: query, schema: { $ref: '#/components/schemas/Discipline' } }
        - { name: organizerRef, in: query, allowReserved: true, schema: { type: string, maxLength: 64 } }
        - { name: from, in: query, schema: { type: string, format: date } }
        - { name: to, in: query, schema: { type: string, format: date } }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Limit' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: '#/components/schemas/Paged'
                          - type: object
                            properties:
                              items: { type: array, items: { $ref: '#/components/schemas/Tournament' } }
        default: { $ref: '#/components/responses/Error' }
  /v1/tournaments/{id}:
    parameters: [{ $ref: '#/components/parameters/TournamentId' }]
    get:
      tags: [tournament]
      summary: Chi tiết giải + tiến độ
      responses:
        '200': { $ref: '#/components/responses/TournamentDetail' }
        default: { $ref: '#/components/responses/Error' }
    patch:
      tags: [tournament]
      summary: Sửa giải (draft / open; đã có người đăng ký thì khoá nội dung / giới / luật điểm / điều kiện trình)
      parameters: [{ $ref: '#/components/parameters/IfMatch' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TournamentPatch' }
      responses:
        '200': { $ref: '#/components/responses/TournamentDetail' }
        default: { $ref: '#/components/responses/Error' }
...
  /v1/tournaments/{id}/entries:
    parameters: [{ $ref: '#/components/parameters/TournamentId' }]
    get:
      tags: [tournament]
      summary: Danh sách đăng ký (kèm điểm xét + cờ trình)
      responses:
        '200': { $ref: '#/components/responses/EntryList' }
        default: { $ref: '#/components/responses/Error' }
    post:
      tags: [tournament]
      summary: Đăng ký người chơi (cặp cố định thì kèm đồng đội)
      description: 'Scope: tournament:operate. Chưa có điểm → 422 NEEDS_ASSESSMENT; sai điều kiện → 422 NOT_ELIGIBLE; hết chỗ → danh sách chờ.'
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [playerId]
              properties:
                playerId: { type: string, format: uuid }
                partnerPlayerId: { type: string, format: uuid }
      responses:
        '201': { $ref: '#/components/responses/EntryList' }
        default: { $ref: '#/components/responses/Error' }
  /v1/tournaments/{id}/entries/{entryId}:
    parameters:
      - { $ref: '#/components/parameters/TournamentId' }
      - { name: entryId, in: path, required: true, schema: { type: string, format: uuid } }
    delete:
      tags: [tournament]
      summary: Rút (sau bốc thăm cả đội rút, trận chưa đánh thành W.O.)
      responses:
        '200': { $ref: '#/components/responses/EntryList' }
        default: { $ref: '#/components/responses/Error' }
  /v1/tournaments/{id}/draw/preview:
    parameters: [{ $ref: '#/components/parameters/TournamentId' }]
    post:
      tags: [tournament]
      summary: Xem trước bốc thăm — không lưu
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                seed: { type: string, maxLength: 64 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/DrawProposal' }
        default: { $ref: '#/components/responses/Error' }
...
  /v1/tournaments/{id}/teams:
    parameters: [{ $ref: '#/components/parameters/TournamentId' }]
    get:
      tags: [tournament]
      summary: Các đội sau bốc thăm
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [items]
                        properties:
                          items: { type: array, items: { $ref: '#/components/schemas/Team' } }
        default: { $ref: '#/components/responses/Error' }
  /v1/tournaments/{id}/matches:
    parameters: [{ $ref: '#/components/parameters/TournamentId' }]
    get:
      tags: [tournament]
      summary: Lịch thi đấu (theo lượt / bảng / sơ đồ)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [items]
                        properties:
                          items: { type: array, items: { $ref: '#/components/schemas/MatchView' } }
        default: { $ref: '#/components/responses/Error' }
    post:
      tags: [tournament]
      summary: Thêm trận tay (play-off, tranh hạng…)
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [teamAId, teamBId]
              properties:
                teamAId: { type: string, format: uuid }
                teamBId: { type: string, format: uuid }
                label: { type: string, maxLength: 60 }
      responses:
        '201': { $ref: '#/components/responses/Match' }
        default: { $ref: '#/components/responses/Error' }
  /v1/tournaments/{id}/standings:
    parameters: [{ $ref: '#/components/parameters/TournamentId' }]
    get:
      tags: [tournament]
      summary: Bảng xếp hạng từng bảng
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [groups]
                        properties:
                          groups:
                            type: array
                            items:
                              type: object
                              required: [groupNo, rows]
                              properties:
                                groupNo: { type: integer }
                                rows: { type: array, items: { $ref: '#/components/schemas/StandingRow' } }
        default: { $ref: '#/components/responses/Error' }
  /v1/tournaments/{id}/bracket:
    parameters: [{ $ref: '#/components/parameters/TournamentId' }]
    get:
      tags: [tournament]
      summary: Sơ đồ loại trực tiếp theo vòng
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [rounds]
                        properties:
                          rounds:
                            type: array
                            items:
                              type: object
                              required: [roundNo, matches]
                              properties:
                                roundNo: { type: integer }
                                matches: { type: array, items: { $ref: '#/components/schemas/MatchView' } }
        default: { $ref: '#/components/responses/Error' }
  /v1/tournaments/{id}/knockout/preview:
    parameters: [{ $ref: '#/components/parameters/TournamentId' }]
    post:
      tags: [tournament]
      summary: Xem trước sơ đồ loại trực tiếp từ kết quả vòng bảng
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [size, positions, teams, firstRound, constraintLevel]
                        properties:
                          size: { type: integer }
                          positions: { type: array, items: { type: string, nullable: true } }
                          teams: { type: object, additionalProperties: { $ref: '#/components/schemas/Team' } }
                          firstRound: { type: array, items: { type: object } }
                          constraintLevel: { type: integer, nullable: true }
        default: { $ref: '#/components/responses/Error' }
...
  /v1/tournaments/{id}/finalize-preview:
    parameters: [{ $ref: '#/components/parameters/TournamentId' }]
    get:
      tags: [tournament]
      summary: Xem trước khi chốt — thứ hạng, điểm trình trước → sau, điểm BXH thành tích
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: '#/components/schemas/FinalizeResult'
                          - type: object
                            required: [ready, pendingMatches, needsKnockout]
                            properties:
                              ready: { type: boolean }
                              pendingMatches: { type: integer }
                              needsKnockout: { type: boolean }
        default: { $ref: '#/components/responses/Error' }
  /v1/tournaments/{id}/finalize:
    parameters: [{ $ref: '#/components/parameters/TournamentId' }]
    post:
      tags: [tournament]
      summary: Chốt giải — áp điểm trình, điểm BXH, thống kê trong một transaction
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      responses:
        '200':
          description: Đã chốt
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        allOf:
                          - $ref: '#/components/schemas/FinalizeResult'
                          - type: object
                            required: [tournament]
                            properties:
                              tournament: { $ref: '#/components/schemas/Tournament' }
        default: { $ref: '#/components/responses/Error' }
  /v1/tournaments/{id}/unfinalize:
    parameters: [{ $ref: '#/components/parameters/TournamentId' }]
    post:
      tags: [tournament]
      summary: Huỷ chốt (chỉ khi chưa ai có thay đổi điểm sau giải)
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      responses:
        '200':
          description: Đã huỷ chốt
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [tournament, rolledBack]
                        properties:
                          tournament: { $ref: '#/components/schemas/Tournament' }
                          rolledBack: { type: array, items: { type: object } }
        default: { $ref: '#/components/responses/Error' }
  /v1/tournaments/{id}/placements:
    parameters: [{ $ref: '#/components/parameters/TournamentId' }]
    get:
      tags: [tournament]
      summary: Thứ hạng chung cuộc (sau khi chốt)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [items]
                        properties:
                          items: { type: array, items: { $ref: '#/components/schemas/Placement' } }
        default: { $ref: '#/components/responses/Error' }
  /v1/tournaments/{id}/stream:
    parameters: [{ $ref: '#/components/parameters/TournamentId' }]
    get:
      tags: [tournament]
      summary: Luồng SSE — tỉ số trực tiếp các trận đang đánh, báo lịch / bảng đấu đổi
      description: |
        text/event-stream. Sự kiện: `snapshot` khi vừa kết nối ({ serverTime, matches: [{ matchId, courtRef, live }] } —
        trận đang đánh đã có người bấm điểm); `score` mỗi lần bấm / hoàn tác ({ matchId, courtRef, live: MatchLive });
        `board` khi sân / trận / danh sách đổi ({ reason } — client tải lại dữ liệu); `ping` 25 giây một lần.
        Service đóng luồng khi token hết hạn; EventSource tự nối lại (gateway ký token mới).
      responses:
        '200':
          description: Luồng sự kiện — mở tới khi client ngắt hoặc token hết hạn
          content:
            text/event-stream:
              schema: { type: string }
        default: { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    serviceToken:
      type: http
      scheme: bearer
      bearerFormat: JWT (ES256)
    eventSignature:
      type: apiKey
      in: header
      name: X-Signature
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Cùng key + cùng body → trả lại đúng response cũ; cùng key + khác body → 422.
      schema: { type: string, pattern: '^[A-Za-z0-9_-]{8,128}$' }
...
    TournamentId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    IfMatch:
      name: If-Match
      in: header
      required: false
      description: 'Version hiện tại (ETag). Ghi kết quả trận bắt buộc có — thiếu → 412, lệch → 409.'
      schema: { type: string, pattern: '^"?[0-9]+"?$' }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100 }
  responses:
    Error:
      description: Lỗi (envelope có `code`)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    Match:
      description: Trận
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Envelope'
              - type: object
                properties:
                  data: { $ref: '#/components/schemas/MatchView' }
...
    TournamentDetail:
      description: Giải (kèm tiến độ)
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Envelope'
              - type: object
                properties:
                  data: { $ref: '#/components/schemas/Tournament' }
    EntryList:
      description: Danh sách đăng ký
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Envelope'
              - type: object
                properties:
                  data:
                    type: object
                    required: [items]
                    properties:
                      items: { type: array, items: { $ref: '#/components/schemas/Entry' } }
...
  schemas:
    Envelope:
      type: object
      required: [success, data, message, errors]
      properties:
        success: { type: boolean }
        data: { description: 'Dữ liệu trả về (null khi không có)' }
        message: { type: string, nullable: true }
        errors: { type: array, nullable: true, items: { $ref: '#/components/schemas/FieldError' } }
        code: { type: string }
...
    Paged:
      type: object
      required: [items, total, page, totalPages]
      properties:
        items: { type: array, items: {} }
        total: { type: integer }
        page: { type: integer }
        totalPages: { type: integer }
...
    Discipline: { type: string, enum: [singles, doubles] }
...
    TournamentPatch:
      type: object
      additionalProperties: false
      properties: &tournamentProps
        name: { type: string, minLength: 3, maxLength: 120 }
        description: { type: string, nullable: true, maxLength: 5000 }
        startsOn: { type: string, format: date }
        tier: { type: string, enum: [club, open, chain] }
        discipline: { $ref: '#/components/schemas/Discipline' }
        genderRule: { type: string, enum: [open, men, women, mixed] }
        pairingMode: { type: string, enum: [fixed, random_balanced] }
        maxPartnerGap: { type: number, nullable: true, minimum: 0, maximum: 6 }
        maxEntries: { type: integer, nullable: true, minimum: 2, maximum: 512 }
        ratingRule: { $ref: '#/components/schemas/RatingRule' }
        format: { type: string, enum: [round_robin, groups_knockout, knockout] }
        groupCount: { type: integer, nullable: true, minimum: 1, maximum: 64 }
        groupMode: { type: string, enum: [seeded, level] }
        advancePerGroup: { type: integer, nullable: true, enum: [1, 2, null] }
        thirdPlaceMatch: { type: boolean }
        scoring: { $ref: '#/components/schemas/Scoring' }
        courtCount: { type: integer, minimum: 1, maximum: 64 }
        matchMinutes: { type: integer, minimum: 5, maximum: 240 }
        rated: { type: boolean }
        ranked: { type: boolean }
    TournamentInput:
      type: object
      additionalProperties: false
      required: [organizerRef, name, startsOn, tier, discipline, genderRule, format, scoring]
      properties:
        <<: *tournamentProps
        organizerRef: { type: string, minLength: 1, maxLength: 64 }
    Tournament:
      type: object
      required: [id, organizerRef, name, startsOn, tier, discipline, genderRule, pairingMode, format, scoring, status, version]
      properties:
        id: { type: string, format: uuid }
        organizerRef: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        startsOn: { type: string, format: date }
        tier: { type: string, enum: [club, open, chain] }
        discipline: { $ref: '#/components/schemas/Discipline' }
        genderRule: { type: string, enum: [open, men, women, mixed] }
        pairingMode: { type: string, enum: [fixed, random_balanced] }
        maxPartnerGap: { type: number, nullable: true }
        maxEntries: { type: integer, nullable: true }
        ratingRule: { $ref: '#/components/schemas/RatingRule' }
        format: { type: string, enum: [round_robin, groups_knockout, knockout] }
        groupCount: { type: integer, nullable: true }
        groupMode: { type: string, enum: [seeded, level] }
        advancePerGroup: { type: integer, nullable: true }
        thirdPlaceMatch: { type: boolean }
        scoring: { $ref: '#/components/schemas/ScoringObject' }
        courtCount: { type: integer }
        matchMinutes: { type: integer }
        rated: { type: boolean }
        ranked: { type: boolean }
        drawSeed: { type: string, nullable: true }
        status: { type: string, enum: [draft, open, drawn, in_progress, finalized, cancelled] }
        stage: { type: string, nullable: true, enum: [group, knockout, null] }
        finalizedAt: { type: string, format: date-time, nullable: true }
        createdByRef: { type: string }
        version: { type: integer }
        createdAt: { type: string, format: date-time, nullable: true }
        progress:
          type: object
          properties:
            entries:
              type: object
              properties:
                registered: { type: integer }
                waitlisted: { type: integer }
                withdrawn: { type: integer }
            matches:
              type: object
              properties:
                total: { type: integer }
                completed: { type: integer }
...
    Team:
      type: object
      required: [id, players, teamRating, groupNo, seed, withdrawn]
      properties:
        id: { type: string, format: uuid }
        players:
          type: array
          items:
            type: object
            required: [id, name, gender]
            properties:
              id: { type: string, format: uuid }
              name: { type: string, nullable: true }
              gender: { type: string, nullable: true }
        teamRating: { type: number }
        groupNo: { type: integer, nullable: true }
        seed: { type: integer, nullable: true }
        withdrawn: { type: boolean }
        groupRank: { type: integer }
...
    StandingRow:
      type: object
      required: [teamId, rank, played, wins, losses, gamesWon, gamesLost, gameDiff, pointsWon, pointsLost, pointDiff, team]
      properties:
        teamId: { type: string, format: uuid }
        rank: { type: integer }
        played: { type: integer }
        wins: { type: integer }
        losses: { type: integer }
        gamesWon: { type: integer }
        gamesLost: { type: integer }
        gameDiff: { type: integer }
        pointsWon: { type: integer }
        pointsLost: { type: integer }
        pointDiff: { type: integer }
        team: { $ref: '#/components/schemas/Team' }
...